@@ -1,6 +1,6 @@
 {
   "name": "@igo2/auth",
-  "version": "1.12.1",
+  "version": "1.13.0-alpha.0",
   "description": "IGO Library",
   "author": "MSP",
   "keywords": ["igo"],
@@ -17,18 +17,18 @@
     "jwt-decode": "^2.2.0",
     "ts-cacheable": "^1.0.5",
     "ts-md5": "^1.2.4",
-    "tslib": "^2.0.0"
+    "tslib": "^2.4.0"
   },
   "peerDependencies": {
-    "@angular/cdk": "^12.2.9",
-    "@angular/common": "^12.2.9",
-    "@angular/core": "^12.2.9",
-    "@angular/forms": "^12.2.9",
-    "@angular/material": "^12.2.9",
-    "@igo2/core": "^1.12.1",
-    "@igo2/utils": "^1.12.1",
-    "rxjs": "^6.6.7",
-    "@azure/msal-angular": "^2.0.0",
-    "@azure/msal-browser": "^2.14.2"
+    "@angular/cdk": "^14.1.2",
+    "@angular/common": "^14.1.2",
+    "@angular/core": "^14.1.2",
+    "@angular/forms": "^14.1.2",
+    "@angular/material": "^14.1.2",
+    "@igo2/core": "^1.13.0-alpha.0",
+    "@igo2/utils": "^1.13.0-alpha.0",
+    "rxjs": "^7.5.6",
+    "@azure/msal-angular": "^2.4.1",
+    "@azure/msal-browser": "^2.28.1"
   }
 }
