@@ -1,6 +1,6 @@
 {
   "name": "@igo2/context",
-  "version": "1.12.1",
+  "version": "1.13.0-alpha.0",
   "description": "IGO Library",
   "author": "MSP",
   "keywords": ["igo"],
@@ -14,20 +14,20 @@
     "url": "https://github.com/infra-geo-ouverte/igo2-lib/issues"
   },
   "peerDependencies": {
-    "@angular/animations": "^12.2.9",
-    "@angular/common": "^12.2.9",
-    "@angular/core": "^12.2.9",
-    "@angular/forms": "^12.2.9",
-    "@angular/material": "^12.2.9",
-    "@igo2/auth": "^1.12.1",
-    "@igo2/common": "^1.12.1",
-    "@igo2/core": "^1.12.1",
-    "@igo2/utils": "^1.12.1",
-    "@igo2/geo": "^1.12.1",
-    "ol": "^6.8.1",
-    "rxjs": "^6.6.7"
+    "@angular/animations": "^14.1.2",
+    "@angular/common": "^14.1.2",
+    "@angular/core": "^14.1.2",
+    "@angular/forms": "^14.1.2",
+    "@angular/material": "^14.1.2",
+    "@igo2/auth": "^1.13.0-alpha.0",
+    "@igo2/common": "^1.13.0-alpha.0",
+    "@igo2/core": "^1.13.0-alpha.0",
+    "@igo2/utils": "^1.13.0-alpha.0",
+    "@igo2/geo": "^1.13.0-alpha.0",
+    "ol": "^6.15.1",
+    "rxjs": "^7.5.6"
   },
   "dependencies": {
-    "tslib": "^2.0.0"
+    "tslib": "^2.4.0"
   }
 }
