@@ -1,6 +1,6 @@
 {
   "name": "@igo2/geo",
-  "version": "1.12.1",
+  "version": "1.13.0-alpha.0",
   "description": "IGO Library",
   "author": "MSP",
   "keywords": ["igo"],
@@ -19,31 +19,31 @@
     "@turf/helpers": "^6.5.0",
     "@turf/line-intersect": "^6.5.0",
     "@turf/point-on-feature": "^6.5.0",
-    "tslib": "^2.0.0",
+    "tslib": "^2.4.0",
     "nosleep.js": "^0.12.0",
     "windows-1252": "^3.0.4"
   },
   "peerDependencies": {
-    "@angular/animations": "^12.2.9",
-    "@angular/common": "^12.2.9",
-    "@angular/core": "^12.2.9",
-    "@angular/forms": "^12.2.9",
-    "@angular/material": "^12.2.9",
-    "@igo2/common": "^1.12.1",
-    "@igo2/core": "^1.12.1",
-    "@igo2/utils": "^1.12.1",
-    "ol": "^6.8.1",
-    "proj4": "^2.7.5",
-    "rxjs": "^6.6.7",
-    "ol-mapbox-style": "^6.4.1",
+    "@angular/animations": "^14.1.2",
+    "@angular/common": "^14.1.2",
+    "@angular/core": "^14.1.2",
+    "@angular/forms": "^14.1.2",
+    "@angular/material": "^14.1.2",
+    "@igo2/common": "^1.13.0-alpha.0",
+    "@igo2/core": "^1.13.0-alpha.0",
+    "@igo2/utils": "^1.13.0-alpha.0",
+    "ol": "^6.15.1",
+    "proj4": "^2.8.0",
+    "rxjs": "^7.5.6",
+    "ol-mapbox-style": "^8.2.1",
     "striptags": "^3.2.0"
   },
   "optionalDependencies": {
     "file-saver": "^2.0.2",
-    "html2canvas": "^1.0.0-rc.5",
-    "jspdf": "^2.3.1",
-    "jszip": "^3.1.5",
-    "moment": "^2.24.0",
+    "html2canvas": "^1.4.1",
+    "jspdf": "^2.5.1",
+    "jszip": "^3.10.1",
+    "moment": "^2.29.4",
     "stream": "^0.0.2"
   }
 }
