@@ -1,6 +1,6 @@
 {
   "name": "@igo2/integration",
-  "version": "1.12.1",
+  "version": "1.13.0-alpha.0",
   "description": "IGO Library",
   "author": "MSP",
   "keywords": ["igo"],
@@ -14,13 +14,13 @@
     "url": "https://github.com/infra-geo-ouverte/igo2-lib/issues"
   },
   "peerDependencies": {
-    "@angular/common": "^12.2.9",
-    "@angular/core": "^12.2.9",
-    "@igo2/geo": "1.12.1",
-    "@igo2/context": "1.12.1",
-    "rxjs": "^6.6.7"
+    "@angular/common": "^14.1.2",
+    "@angular/core": "^14.1.2",
+    "@igo2/geo": "1.13.0-alpha.0",
+    "@igo2/context": "1.13.0-alpha.0",
+    "rxjs": "^7.5.6"
   },
   "dependencies": {
-    "tslib": "^2.0.0"
+    "tslib": "^2.4.0"
   }
 }
