@@ -157,12 +157,25 @@ gulp.task('prepublishOnly', done => {
 
 // ==========================================================
 
+// https://github.com/stylus/stylus/pull/2538
+gulp.task('libs:fixUseAngularMaterialCDK', done => {
+  gulp.src(['./node_modules/@angular/material/core/focus-indicators/_private.scss']).pipe(replace(`@use '@angular/cdk';`, `@use './node_modules/@angular/cdk/_index.scss' as cdk;`)).pipe(gulp.dest('./node_modules/@angular/material/core/focus-indicators/'));
+  gulp.src(['./node_modules/@angular/material/core/style/_menu-common.scss']).pipe(replace(`@use '@angular/cdk';`, `@use './node_modules/@angular/cdk/_index.scss' as cdk;`)).pipe(gulp.dest('./node_modules/@angular/material/core/style/'));
+  gulp.src(['./node_modules/@angular/material/core/_core.scss']).pipe(replace(`@use '@angular/cdk';`, `@use './node_modules/@angular/cdk/_index.scss' as cdk;`)).pipe(gulp.dest('./node_modules/@angular/material/core/'));
+  gulp.src(['./node_modules/@angular/material/core/ripple/_ripple.scss']).pipe(replace(`@use '@angular/cdk';`, `@use './node_modules/@angular/cdk/_index.scss' as cdk;`)).pipe(gulp.dest('./node_modules/@angular/material/core/ripple'));
+  gulp.src(['./node_modules/@angular/material/badge/_badge-theme.scss']).pipe(replace(`@use '@angular/cdk';`, `@use './node_modules/@angular/cdk/_index.scss' as cdk;`)).pipe(gulp.dest('./node_modules/@angular/material/badge'));
+
+  done();
+});
+
+
+
 gulp.task('core:concatStyles', done => {
   gulp
     .src([
       './packages/core/src/style/setup.scss',
-      './packages/core/src/style/all.theming.scss',
       './packages/core/src/style/typography.scss',
+      './packages/core/src/style/all.theming.scss',
       './packages/core/src/style/foreground.scss',
       './packages/core/src/style/theming.scss',
       './packages/core/src/style/core.theming.scss',
@@ -215,7 +228,7 @@ gulp.task('core:concatStyles', done => {
     .pipe(exec(
       'node ./node_modules/scss-bundle/dist/cli/main.js -p ./ -e ./packages/core/src/style/index.theming.scss -o ./dist/core/style/index.theming.scss'
     ))
-    .pipe(wait(500)).on('end', function() { 
+    .pipe(wait(500)).on('end', function() {
       del(['packages/core/src/style/index.theming.scss'], { force: true });
     })
 
@@ -301,6 +314,14 @@ gulp.task('core:copyBundleLocale', done => {
   done();
 });
 
+gulp.task('copyBundleLocaleToDemo', done => {
+  gulp
+    .src('./dist/core/locale/@(fr|en).json', { "allowEmpty": true })
+    .pipe(gulp.dest('./demo/src/locale/libs_locale'));
+
+  done(); //
+}); //
+
 gulp.task('sleep', done => {
   setTimeout(() => done(), 1000);
 });
@@ -310,7 +331,8 @@ gulp.task(
   gulp.series(
     gulp.parallel(['core:bundleLocale.fr', 'core:bundleLocale.en']),
     'sleep',
-    'core:copyBundleLocale'
+    'core:copyBundleLocale',
+    'copyBundleLocaleToDemo'
   )
 );
 
@@ -475,7 +497,7 @@ exports.filename = null;`
 });
 
 
-gulp.task('fixLibs', gulp.parallel(['libs:fixStylus']));
+gulp.task('fixLibs', gulp.parallel(['libs:fixStylus', 'libs:fixUseAngularMaterialCDK']));
 
 // ==========================================================
 
